Added Ability To Open Doors
Added ability to open doors, along with a door to test it. Add the DoorLink script onto an object - when the object is hit with a ray, it will destroy the object whose name is defined in the script. The Door Name can be edited from the object with DoorLink attatched.